Add Selectionjs change Offersjs Change UserCard Adjustg Mockoffers
