+++ mockoffers.json
@@ -4,117 +4,120 @@
     "index": 0,
     "picture": "https://2yrh403fk8vd1hz9ro2n46dd-wpengine.netdna-ssl.com/wp-content/uploads/2017/07/2018-bmw-r-1200-gs-adventure-updates-1.jpg",
     "name": "Head Hensley",
-    "address": "Berlin",
-    "Brand": "BMW",
+    "city": "Berlin",
+    "brand": "BMW",
     "engine": 800,
-    "price": "$168"
+    "price": 16,
+    "email": "finleyrobbins@gspot.com"
   },
   {
     "_id": "5da6d5761dfb8c20a34b8d4f",
     "index": 1,
     "picture": "https://www.motorradhandel.ch/neumotorrad/harley-davidson-xl-1200-ns-sportster-iron/custom/harley-davidson-zentral-schweiz-huenenberg/4838754/12/1/2/",
-    "age": 29,
     "name": "Langley Osborne",
-    "address": "Berlin",
-    "Brand": "Harley Davidson",
+    "city": "Berlin",
+    "brand": "Harley Davidson",
     "engine": 800,
-    "price": "$158"
+    "price": 15,
+    "email": "finleyrobbins@gspot.com"
   },
   {
     "_id": "5da6d5768d12016d19e6e4a4",
     "index": 2,
     "picture": "https://images.ctfassets.net/x7j9qwvpvr5s/4ZumC27QtOqY6a4csG8mWs/bb1cff8cb3d1b7c8bf3f3e840e012d26/Multistrada-1200-Enduro-MY18-Red-01-Model-Preview-1050x650.png",
     "name": "Claudia Cantu",
-    "address": "Hamburg",
-    "Brand": "Ducati",
+    "city": "Hamburg",
+    "brand": "Ducati",
     "engine": 1000,
-    "price": "$160"
+    "price": 160,
+    "email": "finleyrobbins@gspot.com"
   },
   {
     "_id": "5da6d576c4a8b60323805cf2",
     "index": 3,
     "picture": "https://www.motorradhandel.ch/neumotorrad/harley-davidson-xl-1200-ns-sportster-iron/custom/harley-davidson-zentral-schweiz-huenenberg/4838754/12/1/2/",
     "name": "Norris Monroe",
-    "address": "Hamburg",
-    "Brand": "Harley Davidson",
+    "city": "Hamburg",
+    "brand": "Harley Davidson",
     "engine": 950,
-    "price": "$120"
+    "price": 120,
+    "email": "finleyrobbins@gspot.com"
   },
   {
     "_id": "5da6d576fa37a05c7aacc1ce",
     "index": 4,
     "picture": "https://2yrh403fk8vd1hz9ro2n46dd-wpengine.netdna-ssl.com/wp-content/uploads/2017/07/2018-bmw-r-1200-gs-adventure-updates-1.jpg",
-    "age": 29,
     "name": "Michelle Hamilton",
-    "address": "Cologne",
-    "Brand": "BMW",
+    "city": "Cologne",
+    "brand": "BMW",
     "engine": 650,
-    "price": "$110"
+    "price": 110,
+    "email": "finleyrobbins@gspot.com"
   },
   {
     "_id": "5da6d5763819e050ab7638f3",
     "index": 5,
     "picture": "https://images.ctfassets.net/x7j9qwvpvr5s/4ZumC27QtOqY6a4csG8mWs/bb1cff8cb3d1b7c8bf3f3e840e012d26/Multistrada-1200-Enduro-MY18-Red-01-Model-Preview-1050x650.png",
-    "age": 32,
     "name": "Buckley Riggs",
-    "address": "Cologne",
-    "Brand": "Ducati",
+    "city": "Cologne",
+    "brand": "Ducati",
     "engine": 1000,
-    "price": "$145"
+    "price": 145,
+    "email": "finleyrobbins@gspot.com"
   },
   {
     "_id": "5da6d576267edcb7f7662ee7",
     "index": 6,
     "picture": "https://www.motorradhandel.ch/neumotorrad/harley-davidson-xl-1200-ns-sportster-iron/custom/harley-davidson-zentral-schweiz-huenenberg/4838754/12/1/2/",
-    "age": 31,
     "name": "Finley Robbins",
-    "address": "Cologne",
-    "Brand": "Harley Davidson",
+    "city": "Cologne",
+    "brand": "Harley Davidson",
     "engine": 823,
-    "price": "$133"
+    "price": 133,
+    "email": "finleyrobbins@gspot.com"
   },
   {
     "_id": "5da6d576bd2201b0e05ea834",
     "index": 7,
     "picture": "https://images.ctfassets.net/x7j9qwvpvr5s/4ZumC27QtOqY6a4csG8mWs/bb1cff8cb3d1b7c8bf3f3e840e012d26/Multistrada-1200-Enduro-MY18-Red-01-Model-Preview-1050x650.png",
-    "age": 31,
     "name": "Holden Pitts",
-    "address": "Cologne",
-    "Brand": "Ducati",
+    "city": "Cologne",
+    "brand": "Ducati",
     "engine": 904,
-    "price": "$132"
+    "price": 132,
+    "email": "finleyrobbins@gspot.com"
   },
   {
     "_id": "5da6d576fa865a6cb207ee8e",
     "index": 8,
     "picture": "https://2yrh403fk8vd1hz9ro2n46dd-wpengine.netdna-ssl.com/wp-content/uploads/2017/07/2018-bmw-r-1200-gs-adventure-updates-1.jpg",
-    "age": 24,
     "name": "Whitehead Woods",
-    "address": "Berlin",
-    "Brand": "BMW",
+    "city": "Berlin",
+    "brand": "BMW",
     "engine": 839,
-    "price": "$170"
+    "price": 17,
+    "email": "finleyrobbins@gspot.com"
   },
   {
     "_id": "5da6d57694fcecd7b6c5ccf7",
     "index": 9,
     "picture": "https://2yrh403fk8vd1hz9ro2n46dd-wpengine.netdna-ssl.com/wp-content/uploads/2017/07/2018-bmw-r-1200-gs-adventure-updates-1.jpg",
-    "age": 46,
     "name": "Elizabeth Wooten",
-    "address": "Hamburg",
-    "Brand": "BMW",
+    "city": "Hamburg",
+    "brand": "BMW",
     "engine": 800,
-    "price": "$127"
+    "price": 127,
+    "email": "finleyrobbins@gspot.com"
   },
   {
     "_id": "5da6d5765005230b0416aa6c",
     "index": 10,
     "picture": "https://images.ctfassets.net/x7j9qwvpvr5s/4ZumC27QtOqY6a4csG8mWs/bb1cff8cb3d1b7c8bf3f3e840e012d26/Multistrada-1200-Enduro-MY18-Red-01-Model-Preview-1050x650.png",
-    "age": 50,
     "name": "Ladonna Sharp",
-    "address": "Cologne",
-    "Brand": "Ducati",
+    "city": "Cologne",
+    "brand": "Ducati",
     "engine": 650,
-    "price": "$133"
+    "price": 133,
+    "email": "finleyrobbins@gspot.com"
   }
 ]
