+++ src/App.js
@@ -2,9 +2,10 @@ import React from "react";
 import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
 import styled from "styled-components";
 import Chat from "./pages/Chat";
-import Offers from "./pages/Offers";
+import Selection from "./pages/Selection";
 import Profile from "./pages/Profile";
 import Home from "./pages/Home";
+import Offers from "./pages/Offers";
 const Main = styled.main`
   background-color: #f8f5f5;
   display: flex;
@@ -25,6 +26,7 @@ function App() {
           <Route exact path="/offers" component={Offers} />
           <Route exact path="/chat" component={Chat} />
           <Route exact path="/profile" component={Profile} />
+          <Route exact path="/selection" component={Selection} />
         </Switch>
       </Router>
     </Main>
